Feat: 기존 DetailImg를 DetailImg와 DetailRead 분류
